pyproject.toml
[project]
name = "mcp-server-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "google-adk>=0.1.0",
    "google-genai>=1.10.0",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
]
